added wsjtx_lib

+++ gui_ft8.h
@@ -13,7 +13,7 @@ class message
 	int snr;
 	int correct_bits;
 	double off;
-	double hz0;
+	int hz0;
 	std::string msg;
 };
 
@@ -21,7 +21,7 @@ class gui_ft8
 {
   public:
 	void init(lv_obj_t *o_tab, lv_coord_t x, lv_coord_t y, lv_coord_t w, lv_coord_t h);
-	void add_line(int hh, int min, int sec, int snr, int correct_bits, double off, double hz0, std::string msg);
+	void add_line(int hh, int min, int sec, int snr, int correct_bits, double off, int hz0, std::string msg);
 	void clear();
 	void set_group();
 	void reset();
